separated the local execution test
